Upgrade to Astro v5
