+++ src/data/posts/joining-notion.mdx
